Broke off code block into separate component
